fix loading issue when similar coins exist, move coins to home, fill n/a
