+++ gamestonk_terminal/cryptocurrency/discovery/discovery_controller.py
@@ -24,8 +24,6 @@ from gamestonk_terminal.cryptocurrency.discovery import (
     coinpaprika_view,
     coinmarketcap_view,
 )
-from gamestonk_terminal.cryptocurrency.crypto_controller import CRYPTO_SOURCES
-from gamestonk_terminal.cryptocurrency import cryptocurrency_helpers
 
 
 class DiscoveryController:
@@ -74,7 +72,6 @@ class DiscoveryController:
 
         if session and gtff.USE_PROMPT_TOOLKIT:
             choices: dict = {c: {} for c in self.CHOICES}
-            choices["coins"]["--source"] = {c: {} for c in CRYPTO_SOURCES.keys()}
             choices["cggainers"]["-p"] = {
                 c: {} for c in pycoingecko_model.PERIODS.keys()
             }
@@ -121,10 +118,6 @@ class DiscoveryController:
     def print_help(self):
         """Print help"""
         help_text = """
-Discovery Menu:
-
-Overview:
-    coins             search for coins on CoinGecko, Binance, CoinPaprika
 CoinGecko:
     cgtrending        trending coins on CoinGecko
     cgvoted           most voted coins on CoinGecko
@@ -232,75 +225,6 @@ CoinMarketCap:
         self.queue.insert(0, "reset")
         self.queue.insert(0, "quit")
         self.queue.insert(0, "quit")
-
-    @try_except
-    def call_coins(self, other_args):
-        """Process coins command"""
-        parser = argparse.ArgumentParser(
-            prog="coins",
-            add_help=False,
-            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-            description="""Shows list of coins available on CoinGecko, CoinPaprika and Binance.If you provide name of
-            coin then in result you will see ids of coins with best match for all mentioned services.
-            If you provide ALL keyword in your search query, then all coins will be displayed. To move over coins you
-            can use pagination mechanism with skip, top params. E.g. coins ALL --skip 100 --limit 30 then all coins
-            from 100 to 130 will be displayed. By default skip = 0, limit = 10.
-            If you won't provide source of the data everything will be displayed (CoinGecko, CoinPaprika, Binance).
-            If you want to search only in given source then use --source flag. E.g. if you want to find coin with name
-            uniswap on CoinPaprika then use: coins uniswap --source cp --limit 10
-                """,
-        )
-
-        parser.add_argument(
-            "-c",
-            "--coin",
-            help="Coin you search for",
-            dest="coin",
-            required="-h" not in other_args,
-            type=str,
-        )
-
-        parser.add_argument(
-            "-s",
-            "--skip",
-            default=0,
-            dest="skip",
-            help="Skip n of records",
-            type=check_positive,
-        )
-
-        parser.add_argument(
-            "-l",
-            "--limit",
-            default=10,
-            dest="limit",
-            help="Limit of records",
-            type=check_positive,
-        )
-
-        parser.add_argument(
-            "--source",
-            dest="source",
-            help="Source of data.",
-            type=str,
-            choices=CRYPTO_SOURCES.keys(),
-        )
-
-        if other_args and not other_args[0][0] == "-":
-            other_args.insert(0, "-c")
-
-        ns_parser = parse_known_args_and_warn(
-            parser, other_args, EXPORT_ONLY_RAW_DATA_ALLOWED
-        )
-        if ns_parser:
-            cryptocurrency_helpers.display_all_coins(
-                coin=ns_parser.coin,
-                source=ns_parser.source,
-                top=ns_parser.limit,
-                skip=ns_parser.skip,
-                show_all=bool("ALL" in other_args),
-                export=ns_parser.export,
-            )
 
     @try_except
     def call_cggainers(self, other_args):
